djangoapp/requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
backports.entry-points-selectable==1.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.23
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-timezone-field==7.1
djangorestframework==3.16.0
exceptiongroup==1.3.0
honcho==2.0.0
iniconfig==2.1.0
kafka-python==2.2.15
kombu==5.5.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
